+++ api/httpserver/rest/rewards_test.go
@@ -1,4 +1,4 @@
-package api
+package rest_test
 
 import (
 	"testing"
