+++ collector/collector_test.go
@@ -1,8 +1,10 @@
-package collector
+package collector_test
 
 import (
 	"context"
 	"fmt"
+	testseed2 "github.com/spacemeshos/explorer-backend/test/testseed"
+	"github.com/spacemeshos/explorer-backend/test/testserver"
 	"os"
 	"testing"
 	"time"
@@ -13,15 +15,13 @@ import (
 
 	"github.com/spacemeshos/explorer-backend/collector"
 	"github.com/spacemeshos/explorer-backend/storage"
-	"github.com/spacemeshos/explorer-backend/testhelpers/testseed"
-	"github.com/spacemeshos/explorer-backend/testhelpers/testserver"
 )
 
 const testAPIServiceDB = "explorer_test"
 
 var (
 	dbPort       = 27017
-	generator    *testseed.SeedGenerator
+	generator    *testseed2.SeedGenerator
 	node         *testserver.FakeNode
 	collectorApp *collector.Collector
 	storageDB    *storage.Storage
@@ -49,8 +49,8 @@ func TestMain(m *testing.M) {
 		os.Exit(1)
 	}
 
-	seed := testseed.GetServerSeed()
-	generator = testseed.NewSeedGenerator(seed)
+	seed := testseed2.GetServerSeed()
+	generator = testseed2.NewSeedGenerator(seed)
 	if err = generator.GenerateEpoches(10); err != nil {
 		fmt.Println("failed to generate epochs", err)
 		os.Exit(1)
